requirements.txt
flask
pywebview
pandas
openpyxl
numpy
pyinstaller